+++ R/icarfit.R
@@ -5,11 +5,11 @@
 #' @param nburnin Number of iterations discarded for burn-in. Default is 250,000.
 #' @param thin Thinning interval
 #' @param nchains Number of MCMC chains
-#' @param sigmaPrior A string defining prior for the sigma parameter. Default is 'dexp(1)'. 
+#' @param sigmaPrior A string defining prior for the sigma parameter. Default is 'dexp(rate=1)'. 
 #' @param sigmaSampler A list containing settings for the MCMC sampler. Default is null and employs nimble's Default sampler (RW sampler).
 #' @param parallel Logical specifying whether the chains should be run in parallel or not.
 #' @param seeds Random seed for each chain. Default is 1:4.
-#' @details The function estimates a vector temporally autocorrelated  probability values on the observed data using MCMC as implemented by the nimble package. The model is effectively non-parametric, and at its core it is an implementation of a 1D random ICAR model. Users can specify the prior for the variance parameter through the argument \code{sigmaPrior}. Different settings for this parameter can greatly influence the estimates of the probability vectors. For example using \code{sigmaPrior=dexp(100)} instead of the default \code{sigmaPrior=dexp(1)} would lead to 'flatter' time-series with higher temporal autocorrelation. Please consult the nimble package manual for the syntax required in specifying the prior. MCMC settings such as the choice the sampler, number of iterations, chains, etc can also be specified. Please not that the function is computationally intensive and might require a larger number of iterations to reach satisfactory MCMC convergence.
+#' @details The function estimates a vector temporally autocorrelated  probability values on the observed data using MCMC as implemented by the nimble package. The model is effectively non-parametric, and at its core it is an implementation of a 1D random ICAR model. Users can specify the prior for the variance parameter through the argument \code{sigmaPrior}. Different settings for this parameter can greatly influence the estimates of the probability vectors. For example using \code{sigmaPrior=dexp(100)} instead of the default \code{sigmaPrior=dexp(1)} would lead to 'flatter' time-series with higher temporal autocorrelation. The distribution parameters defined in \code{sigmaPrior} is also used to generate initialisation values for the MCMC. Please consult the nimble package manual for the syntax required in specifying the prior. MCMC settings such as the choice the sampler, number of iterations, chains, etc can also be specified. Please not that the function is computationally intensive and might require a larger number of iterations to reach satisfactory MCMC convergence.
 
 #' @return A \code{fittedICAR} class object containing the original ProbMat class object, posteriors of the probabilities for each time-block and the variance parameter along with their MCMC diagnostics (Gelman Rubin statistic and effective sample size).
 
@@ -20,7 +20,7 @@
 #' @export
 
 
-icarfit  <- function(x,niter=100000,nburnin=50000,thin=10,nchains=4,sigmaPrior='dexp(1)',sigmaSampler=NULL,parallel=FALSE,seeds=1:4)
+icarfit  <- function(x,niter=100000,nburnin=50000,thin=10,nchains=4,sigmaPrior='dexp(rate=1)',sigmaSampler=NULL,parallel=FALSE,seeds=1:4)
 {
 	#Handle cleaning of GlobalEnv on exit
 	envobj <- ls(envir=.GlobalEnv)
@@ -30,6 +30,13 @@ icarfit  <- function(x,niter=100000,nburnin=50000,thin=10,nchains=4,sigmaPrior='
 
 	# Initial Warnings
 	if (nchains==1) {warning('Running MCMC on single chain')}
+
+	# Check prior definitions
+	supported.distributions  <- data.frame(d=c('dnorm','dexp','dbeta','dchisq','dgamma','dlogis','dunif','dt','dweib','dlnorm'),r=c('rnorm','rexp','rbeta','rchisq','rgamma','rlogis','runif','rt','rweibull','rlnorm'))
+	if(!sub("\\(.*","",sigmaPrior)%in%supported.distributions$d){stop(paste0('Unsupported distribution in prior definition. Please use one of the following: ',paste(supported.distributions$d,collapse=', ')))}
+	sigmaPrior.rand1  <- supported.distributions$r[supported.distributions$d==strsplit(sigmaPrior,"\\(")[[1]][1]]
+	sigmaPrior.rand2 <- gsub("\\)","",strsplit(sigmaPrior,"\\(")[[1]][2])
+	sigmaPrior.rand <- paste0(sigmaPrior.rand1,"(n=1,",sigmaPrior.rand2,")")
 
 	# Define Data
 	d  <- list(theta=x$pmat)
@@ -76,14 +83,14 @@ icarfit  <- function(x,niter=100000,nburnin=50000,thin=10,nchains=4,sigmaPrior='
 			lpseq[1:n.tblocks] ~ dcar_normal(adj[1:L], weights[1:L], num[1:n.tblocks], tau, zero_mean = 0)
 			lpseqSS  <- sum(exp(lpseq[1:n.tblocks]))
 			tau <- 1/sigma^2
-			sigma  ~ dexp(1)
+			sigma  ~ sigmaPrior
 		})
-		icarmodel <- gsub('dexp\\(1\\)', sigmaPrior, deparse(icarmodel)) |> parse(text=_)
+		icarmodel <- gsub('sigmaPrior', sigmaPrior, deparse(icarmodel)) |> parse(text=_)
 
 		inits  <- vector('list',length=nchains)
 		for (k in 1:nchains)
 		{
-			inits[[k]]  <- list(sigma=rexp(1),lpseq=rnorm(constants$n.tblocks,0,0.5))
+			inits[[k]]  <- list(sigma=eval(parse(text=sigmaPrior.rand)),lpseq=rnorm(constants$n.tblocks,0,0.5))
 		}
 		message('Compiling nimble model...')
 		suppressMessages(model  <- nimbleModel(icarmodel,constants=constants,data=d,inits=inits[[1]]))
@@ -92,7 +99,6 @@ icarfit  <- function(x,niter=100000,nburnin=50000,thin=10,nchains=4,sigmaPrior='
 		if (!is.null(sigmaSampler))
 		{
 			suppressMessages(conf$removeSamplers('sigma'))
-			# 	sigmaSampler=list('sigma',type='slice')
 			suppressMessages(do.call(conf$addSampler,sigmaSampler))
 		}
 		suppressMessages(conf$addMonitors('p'))
@@ -107,7 +113,7 @@ icarfit  <- function(x,niter=100000,nburnin=50000,thin=10,nchains=4,sigmaPrior='
 	{
 		# Addresses R CMD Check NOTES
 		message('Running in parallel - progress bar will no be visualised')
-		runfun  <- function(seed,constants,d,niter,thin,nburnin,sigmaPrior,sigmaSampler)
+		runfun  <- function(seed,constants,d,niter,thin,nburnin,sigmaPrior,sigmaPrior.rand,sigmaSampler)
 		{
 
 			returnType <- n.tblocks  <- lpseq <- sigma <- nimStop <- nimMatrix <- rAoristicGeneral_vector <- dAOG <- rAOG <- NULL
@@ -137,18 +143,16 @@ icarfit  <- function(x,niter=100000,nburnin=50000,thin=10,nchains=4,sigmaPrior='
 				theta[,] ~ dAOG(p=p[1:n.tblocks])
 				for (i in 1:n.tblocks)
 				{
-# 					p[i]  <- lpseq[i]^2 / lpseqSS
 					p[i]  <- exp(lpseq[i]) / lpseqSS
 				}
 				lpseq[1:n.tblocks] ~ dcar_normal(adj[1:L], weights[1:L], num[1:n.tblocks], tau, zero_mean = 0)
-# 				lpseqSS  <- sum(lpseq[1:n.tblocks]^2)
 				lpseqSS  <- sum(exp(lpseq[1:n.tblocks]))
 				tau <- 1/sigma^2
-				sigma  ~ dexp(1)
+				sigma  ~ sigmaPrior
 			})
-			icarmodel <- gsub('dexp\\(1\\)', sigmaPrior, deparse(icarmodel)) |> parse(text=_)
+			icarmodel <- gsub('sigmaPrior', sigmaPrior, deparse(icarmodel)) |> parse(text=_)
 			set.seed(seed)
-			inits  <- list(sigma=rexp(1),lpseq=rnorm(constants$n.tblocks,0,0.5))
+			inits  <- list(sigma=eval(parse(text=sigmaPrior.rand)),lpseq=rnorm(constants$n.tblocks,0,0.5))
 			model  <- nimbleModel(icarmodel,constants=constants,data=d,inits=inits)
 			cModel <- compileNimble(model)
 			conf <- configureMCMC(model)
@@ -166,7 +170,7 @@ icarfit  <- function(x,niter=100000,nburnin=50000,thin=10,nchains=4,sigmaPrior='
 		ncores  <- nchains
 		cl  <- makeCluster(ncores)
 		clusterEvalQ(cl,{library(nimble)})
-		out  <- parLapply(cl=cl,X=seeds,fun=runfun,d=d,constants=constants,nburnin=nburnin,niter=niter,thin=thin,sigmaPrior=sigmaPrior,sigmaSampler=sigmaSampler)
+		out  <- parLapply(cl=cl,X=seeds,fun=runfun,d=d,constants=constants,nburnin=nburnin,niter=niter,thin=thin,sigmaPrior=sigmaPrior,sigmaSampler=sigmaSampler,sigmaPrior.rand=sigmaPrior.rand)
 		stopCluster(cl)
 		results <- out
 	}