+++ NEWS.md
@@ -1,3 +1,10 @@
+# Version 0.2.0
+ * Fixed bug on logistic growth model plot function
+ * Fixed bug on prior definition for exponential growth model
+ * Added NEW double exponential model and related plot function
+ * Added NEW Monte-Carlo simulation function
+ * Updated parameter initialisation routine in all model. Initial values for MCMC are drawn from the prior distribution.
+
 # Version 0.1.4
  - Fixed bug caused by changes in nimble v.1.2.0.
 .